pgw#1518: fold the skeleton _resolve tests into the streaming module — pgw master's `fast gates` is red for every lane
`test_skeleton_pipeline_submodule_pgw1518.py` landed with #1079 (`c24636bb`)
and pgw#849/#1362's naming guard refuses it: a NEW top-level test module named
for an issue. `fast gates` is pgw's ONLY required context, so master is red
and no lane can merge — including a fleet-blocking serve-path fix with three
H100 rentals held behind it.

Zero behaviour change and zero new files: the four tests move into
`test_weight_streaming.py`, which is the same subsystem (both exercise
`serving/streaming/skeleton.py`) and already builds real projected trees full
of pointer stubs. The lineage stays as a one-line comment naming pgw#1518,
which is the trade DESIGN-RULINGS 4.34b makes — the ledger keeps the
narrative, the file name keeps the subject. Their import joins the
skeleton import already at the top of that module rather than adding a second.

Folding rather than renaming, following the se#790 lane's example on the same
guard an hour earlier: renaming satisfies the letter while still adding a
file, and the guard's whole argument is that the corpus grew while everyone
was doing correct, ruled work.

Not my file. Done by the th#2204 lane because master is red for everyone and
the alternative was leaving it that way; the tests and their intent are
untouched, and the owning lane is told.
